TST/REF: collect Index setops tests
