+++ pandas/tests/indexes/test_numeric.py
@@ -1,4 +1,4 @@
-from datetime import datetime, timedelta
+from datetime import datetime
 
 import numpy as np
 import pytest
@@ -408,18 +408,6 @@ class NumericInt(Numeric):
 
         assert not index.astype(dtype=object).identical(index.astype(dtype=self._dtype))
 
-    def test_union_noncomparable(self):
-        # corner case, non-Int64Index
-        index = self.create_index()
-        other = Index([datetime.now() + timedelta(i) for i in range(4)], dtype=object)
-        result = index.union(other)
-        expected = Index(np.concatenate((index, other)))
-        tm.assert_index_equal(result, expected)
-
-        result = other.union(index)
-        expected = Index(np.concatenate((other, index)))
-        tm.assert_index_equal(result, expected)
-
     def test_cant_or_shouldnt_cast(self):
         msg = (
             "String dtype not supported, "
@@ -535,19 +523,6 @@ class TestInt64Index(NumericInt):
         arr = Index([1, 2, 3, 4], dtype=object)
         assert isinstance(arr, Index)
 
-    def test_intersection(self):
-        index = self.create_index()
-        other = Index([1, 2, 3, 4, 5])
-        result = index.intersection(other)
-        expected = Index(np.sort(np.intersect1d(index.values, other.values)))
-        tm.assert_index_equal(result, expected)
-
-        result = other.intersection(index)
-        expected = Index(
-            np.sort(np.asarray(np.intersect1d(index.values, other.values)))
-        )
-        tm.assert_index_equal(result, expected)
-
 
 class TestUInt64Index(NumericInt):
 
@@ -564,14 +539,8 @@ class TestUInt64Index(NumericInt):
     def index(self, request):
         return UInt64Index(request.param)
 
-    @pytest.fixture
-    def index_large(self):
-        # large values used in TestUInt64Index where no compat needed with Int64/Float64
-        large = [2 ** 63, 2 ** 63 + 10, 2 ** 63 + 15, 2 ** 63 + 20, 2 ** 63 + 25]
-        return UInt64Index(large)
-
     def create_index(self) -> UInt64Index:
-        # compat with shared Int64/Float64 tests; use index_large for UInt64 only tests
+        # compat with shared Int64/Float64 tests
         return UInt64Index(np.arange(5, dtype="uint64"))
 
     def test_constructor(self):
@@ -595,44 +564,6 @@ class TestUInt64Index(NumericInt):
         idx = UInt64Index([1, 2 ** 63 + 1], dtype=np.uint64)
         res = Index([1, 2 ** 63 + 1], dtype=np.uint64)
         tm.assert_index_equal(res, idx)
-
-    def test_intersection(self, index_large):
-        other = Index([2 ** 63, 2 ** 63 + 5, 2 ** 63 + 10, 2 ** 63 + 15, 2 ** 63 + 20])
-        result = index_large.intersection(other)
-        expected = Index(np.sort(np.intersect1d(index_large.values, other.values)))
-        tm.assert_index_equal(result, expected)
-
-        result = other.intersection(index_large)
-        expected = Index(
-            np.sort(np.asarray(np.intersect1d(index_large.values, other.values)))
-        )
-        tm.assert_index_equal(result, expected)
-
-
-@pytest.mark.parametrize("dtype", ["int64", "uint64"])
-def test_int_float_union_dtype(dtype):
-    # https://github.com/pandas-dev/pandas/issues/26778
-    # [u]int | float -> float
-    index = Index([0, 2, 3], dtype=dtype)
-    other = Float64Index([0.5, 1.5])
-    expected = Float64Index([0.0, 0.5, 1.5, 2.0, 3.0])
-    result = index.union(other)
-    tm.assert_index_equal(result, expected)
-
-    result = other.union(index)
-    tm.assert_index_equal(result, expected)
-
-
-def test_range_float_union_dtype():
-    # https://github.com/pandas-dev/pandas/issues/26778
-    index = pd.RangeIndex(start=0, stop=3)
-    other = Float64Index([0.5, 1.5])
-    result = index.union(other)
-    expected = Float64Index([0.0, 0.5, 1, 1.5, 2.0])
-    tm.assert_index_equal(result, expected)
-
-    result = other.union(index)
-    tm.assert_index_equal(result, expected)
 
 
 @pytest.mark.parametrize(
@@ -675,15 +606,3 @@ def test_float64_index_equals():
 
     result = string_index.equals(float_index)
     assert result is False
-
-
-def test_float64_index_difference():
-    # https://github.com/pandas-dev/pandas/issues/35217
-    float_index = Index([1.0, 2, 3])
-    string_index = Index(["1", "2", "3"])
-
-    result = float_index.difference(string_index)
-    tm.assert_index_equal(result, float_index)
-
-    result = string_index.difference(float_index)
-    tm.assert_index_equal(result, string_index)
